MNIST/requirements.txt
tensorflow-gpu==2.10.1
pandas
numpy
matplotlib